Merge branch 'master' of https://github.com/CIRB/django-fixmystreet

Conflicts:
	django_fixmystreet/backoffice/templates/pro/reports/show.html

+++ django_fixmystreet/backoffice/views/groups.py
@@ -8,6 +8,8 @@ from django.contrib import messages
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext as _
 from django.db import IntegrityError
+from django.core.exceptions import PermissionDenied
+
 
 from django_fixmystreet.backoffice.forms import GroupForm
 from django_fixmystreet.fixmystreet.models import OrganisationEntity, UserOrganisationMembership, FMSUser
@@ -20,12 +22,11 @@ def list_groups(request):
 
     group_types = [t[0] for t in OrganisationEntity.ENTITY_TYPE_GROUP]
     groups = OrganisationEntity.objects.filter(type__in=group_types)
-    groups = groups.filter(dependency=request.fmsuser.organisation)
 
-    #~ if current_user.organisation:
-        #~ groups = groups.filter(dependency=current_user.organisation)
-    #~ elif not request.user.is_superuser:
-        #~ raise PermissionDenied()
+    if current_user.organisation:
+        groups = groups.filter(dependency=request.fmsuser.organisation)
+    elif not request.user.is_superuser:
+        raise PermissionDenied()
 
     return render_to_response("pro/auth/groups_list.html", {
         'groups': groups,
@@ -100,7 +101,9 @@ def add_membership(request, group_id, user_id):
         organisation = OrganisationEntity.objects.get(id=group_id)
         user = FMSUser.objects.get(id=user_id)
 
-        user_organisation_membership = UserOrganisationMembership(user=user, organisation=organisation)
+        contact = not organisation.memberships.exists()
+        user_organisation_membership = organisation.memberships.create(user=user, contact_user=contact)
+        # user_organisation_membership = UserOrganisationMembership(user=user, organisation=organisation)
 
         try:
             user_organisation_membership.save()
@@ -122,6 +125,24 @@ def remove_membership(request, membership_id):
         return HttpResponse('Permission Denied')
 
     user_organisation_membership = UserOrganisationMembership.objects.get(id=membership_id)
+
+    if user_organisation_membership.contact_user:
+        return HttpResponse('Can not remove the contact user')
+
     user_organisation_membership.delete()
+
+    return HttpResponse('OK')
+
+
+def contact_membership(request, membership_id):
+    can_edit = request.fmsuser.leader
+
+    if not can_edit:
+        return HttpResponse('Permission Denied')
+
+    user_organisation_membership = UserOrganisationMembership.objects.get(id=membership_id)
+    user_organisation_membership.organisation.memberships.all().update(contact_user=False)
+    user_organisation_membership.contact_user = True
+    user_organisation_membership.save()
 
     return HttpResponse('OK')
